[MIGRATIONS] More sanity checks

+++ core/migrations/Migration20141202211549PlgMembersDashboard.php
@@ -18,10 +18,13 @@ class Migration20141202211549PlgMembersDashboard extends Base
 		// get dashboard params
 		$pluginParams = $this->getParams('plg_members_dashboard');
 
-		// delete all null user preferences
-		$sql = "DELETE FROM `#__xprofiles_dashboard_preferences` WHERE `preferences`='[]';";
-		$this->db->setQuery($sql);
-		$this->db->query();
+		if ($this->db->tableExists('#__xprofiles_dashboard_preferences'))
+		{
+			// delete all null user preferences
+			$sql = "DELETE FROM `#__xprofiles_dashboard_preferences` WHERE `preferences`='[]';";
+			$this->db->setQuery($sql);
+			$this->db->query();
+		}
 
 		// only continue if plugin defaults are NOT set
 		$defaults = trim($pluginParams->get('defaults', ''));
@@ -30,35 +33,37 @@ class Migration20141202211549PlgMembersDashboard extends Base
 			return;
 		}
 
-		// get top 6 modules
-		$sql = "SELECT id FROM `#__modules` WHERE `position`='memberDashboard' AND `published`=1 AND `client_id`=0 ORDER BY `ordering` LIMIT 6;";
-		$this->db->setQuery($sql);
-		$modules = $this->db->loadColumn();
-
 		// array to hold new defaults
 		$defaults = array();
 
-		// create default
-		$col = 0;
-		$row = 1;
-		foreach ($modules as $k => $module)
+		if ($this->db->tableExists('#__modules'))
 		{
-			$col = $col + 1;
-			if ($col > 3)
+			// get top 6 modules
+			$sql = "SELECT id FROM `#__modules` WHERE `position`='memberDashboard' AND `published`=1 AND `client_id`=0 ORDER BY `ordering` LIMIT 6;";
+			$this->db->setQuery($sql);
+			$modules = $this->db->loadColumn();
+
+			// create default
+			$col = 0;
+			$row = 1;
+			foreach ($modules as $k => $module)
 			{
-				$col = 1;
-				$row = 3;
+				$col = $col + 1;
+				if ($col > 3)
+				{
+					$col = 1;
+					$row = 3;
+				}
+
+				array_push($defaults, array(
+					'module' => $module,
+					'col'    => $col,
+					'row'    => $row,
+					'size_x' => 1,
+					'size_y' => 2
+				));
 			}
-
-			array_push($defaults, array(
-				'module' => $module,
-				'col'    => $col,
-				'row'    => $row,
-				'size_x' => 1,
-				'size_y' => 2
-			));
 		}
-
 		// update params & save
 		$pluginParams->set('defaults', $defaults);
 		$this->savePluginParams('members', 'dashboard', $pluginParams->toArray());
